+++ modules/light-clients/08-wasm/keeper/options_test.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"errors"
 
-	wasmvmtypes "github.com/CosmWasm/wasmvm/v2/types"
+	wasmvmtypes "github.com/CosmWasm/wasmvm/v3/types"
 
 	"github.com/cosmos/cosmos-sdk/runtime"
 	sdk "github.com/cosmos/cosmos-sdk/types"